address review: normalize started_at hint, reclassify peek test
- AnnotationsTable: normalize `started_at` to a numeric ms hint via a
  `toOccurredAtMsHint` helper (handles numeric and ISO-string values,
  drops non-finite) so the hint always satisfies the
  `z.number().int().optional()` query contract and the drawer `t` param.
- Rename the peek test to `.integration.test.tsx` (it renders the
  component and mounts Chakra) and restructure into given/when/it BDD
  blocks per the repo testing conventions.
